fix flake

+++ adaptdl/adaptdl/goodput.py
@@ -116,11 +116,13 @@ class GoodputFunction(object):
                 np.logical_and(num_replicas == 1,
                                local_bsz > self._init_batch_size),
                 np.maximum(accum_steps, 1), accum_steps).astype(int)
-            atomic_bsz = np.ceil(local_bsz / (accum_steps + 1) - eps).astype(int)
+            atomic_bsz = np.ceil(
+                local_bsz / (accum_steps + 1) - eps).astype(int)
         else:
             accum_steps = np.zeros_like(local_bsz, dtype=np.int)
-            atomic_bsz = np.where(num_replicas == 1
-                self._init_batch_size, local_bsz)
+            atomic_bsz = np.where(
+                num_replicas == 1,
+                self._init_batch_size, local_bsz).astype(int)
 
         # Evaluate the goodput of all candidate configurations.
         goodput = self.evaluate(num_nodes, num_replicas,
